Moved the runCommand to it’s own class
