+++ src/Skeletor/Manager/InstallerInterface.php
@@ -5,5 +5,5 @@ interface InstallerInterface
 {
     public function prepareFrameworkCommand(string $framework, string $version);
     public function preparePackageCommand(string $package, string $version, string $packageOptions);
-    public function runCommand(string $command);
+    public function updateComposerFile(array $updates);
 }
